Extract PopulateEffectsMenu
